fix: resolve integration test build errors and package inconsistencies
- Fix setupIperfVethPair function availability by adding it to iperf_bandwidth_test.go
- Standardize all integration test packages to integration_test
- Remove duplicate helper file that caused package conflicts
- Fix build errors preventing integration tests from running
