+++ test/integration/sqlite_event_store_test.go
@@ -1,4 +1,4 @@
-package integration
+package integration_test
 
 import (
 	"context"
