+++ test/integration/statistics_integration_test.go
@@ -1,4 +1,4 @@
-package integration
+package integration_test
 
 import (
 	"testing"
